+++ include/MBSBuffer.h
@@ -20,22 +20,24 @@
 #ifndef MBSUTIL_BUFFER_H
 #define MBSUTIL_BUFFER_H
 
+#include <stdio.h>
+
 class MBSBuffer
 {
 public:
 	unsigned char *dados;
 	unsigned char *pntE,*pntL;
 
-	MBSBuffer(unsigned long tam=0);
+	MBSBuffer(size_t tam=0);
 	virtual ~MBSBuffer() { if(dados) delete[] dados; }
 
 	void reset();	//pntL=pntE=dados
-	int mudaTamanho(unsigned long tam);
-	inline unsigned long pegaTamanho() const { return tamanho; }
-	inline unsigned long disponiveis() const { return (pntE>pntL)?(pntE-pntL):0; }
-	inline unsigned long livres() const { return tamanho-(pntE-dados); }
-	inline unsigned long ocupados() const { return pntE-dados; }
-	unsigned long append(MBSBuffer& b, unsigned long qtd=0);
+	int mudaTamanho(size_t tam);
+	inline size_t pegaTamanho() const { return tamanho; }
+	inline size_t disponiveis() const { return (pntE>pntL)?(pntE-pntL):0; }
+	inline size_t livres() const { return tamanho-(pntE-dados); }
+	inline size_t ocupados() const { return pntE-dados; }
+	unsigned long append(MBSBuffer& b, size_t qtd=0);
 		//anexa qtd bytes de b->pntL em pntE
 		//se qtd==0, anexa disponiveis() bytes de b->pntL em pntE
 
@@ -45,10 +47,10 @@ public:
 	int writeShort(unsigned short us);
 	int readLong(unsigned long *ul);
 	int writeLong(unsigned long ul);
-	int readString(char *dest, unsigned long max);
-	int writeString(const char *orig, unsigned long max=0);
+	int readString(char *dest, size_t max);
+	int writeString(const char *orig, size_t max=0);
 protected:
-	unsigned long tamanho;
+	size_t tamanho;
 };
 
 #endif
